refactor(core): ♻️ rework Plugin abstract class definition APIs
BREAKING CHANGE:

- add init, shutdown abstract methods
- add _setRuntime method accessible through the _setRuntime Symbol (meant to act as a C++ friend class)
- move types into dedicated plugin.types file
- renames ExecutorImplementation to Executor
