+++ packages/core/src/runtime/providers/index.ts
@@ -1,3 +1,4 @@
 export * from "./model";
 export * from "./memory";
 export * from "./plugin";
+export * from "./plugin.types";
